+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Types of changes
 
 - `Added` flag `--include` (short `-i`) to only scan/dump a specific list of fields, this flag is repeatable
 - `Added` flag `--alias` (short `-a`) to rename fields on the fly, this flag is repeatable
+- `Added` flag `--watch` (short `-w`) to the dump command
 - `Fixed` self reference link are no longer counted in the links counter while scanning
 
 ## [0.1.0]